# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "🎶 MusicBoxAPI",
        "description": "API RESTful pour la gestion des artistes, albums et chansons.\n *                   Cette API permet de gérer les ressources principales :\n *                   - Artistes\n *                   - Albums\n *                   - Chansons\n *                   Elle inclut des fonctionnalités de recherche, pagination et filtres.",
        "contact": {
            "name": "MusicBox API Team",
            "email": "support@musicboxapi.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Authentifier un utilisateur (login)",
                "description": "Permet à un utilisateur de se connecter et de recevoir un token d'accès.",
                "operationId": "bc76a9d52929cab7a147ca7b5c527430",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "test@test.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Yassine"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "test@test.com"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|XyzAbc12345..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Identifiants invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid credentials"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Déconnexion (logout)",
                "description": "Supprime le token de l’utilisateur actuellement connecté.",
                "operationId": "0ab4e3cf7bc4f7ec81b48f7928c7de7c",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/songs": {
            "get": {
                "tags": [
                    "Songs"
                ],
                "summary": "Lister toutes les chansons",
                "operationId": "0388bb28c35eec49b8b492d169ce3575",
                "responses": {
                    "200": {
                        "description": "Liste paginée des chansons"
                    }
                }
            },
            "post": {
                "tags": [
                    "Songs"
                ],
                "summary": "Créer une nouvelle chanson",
                "operationId": "2f1970b82d91cc715578d972189f19cb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "duration",
                                    "album_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Lose Yourself"
                                    },
                                    "duration": {
                                        "type": "integer",
                                        "example": 326
                                    },
                                    "album_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Chanson créée avec succès"
                    }
                }
            }
        },
        "/api/songs/{id}": {
            "get": {
                "tags": [
                    "Songs"
                ],
                "summary": "Afficher une chanson par ID",
                "operationId": "4325a96361d1d3820cc0580eaae9197c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la chanson",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chanson trouvée"
                    },
                    "404": {
                        "description": "Chanson non trouvée"
                    }
                }
            },
            "put": {
                "tags": [
                    "Songs"
                ],
                "summary": "Mettre à jour une chanson",
                "operationId": "452de7f8f11a121cd11e9430565632c3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la chanson",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Stan"
                                    },
                                    "duration": {
                                        "type": "integer",
                                        "example": 400
                                    },
                                    "album_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Chanson mise à jour avec succès"
                    },
                    "404": {
                        "description": "Chanson non trouvée"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Songs"
                ],
                "summary": "Supprimer une chanson",
                "operationId": "829c1a52ae5013ce3ff4b9829a31485f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la chanson",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Chanson supprimée"
                    },
                    "404": {
                        "description": "Chanson non trouvée"
                    }
                }
            }
        },
        "/api/songs/search": {
            "get": {
                "tags": [
                    "Songs"
                ],
                "summary": "Rechercher une chanson par titre ou artiste",
                "operationId": "f09417af32c117732bb14bc6def5bbdc",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Titre partiel ou complet de la chanson",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Lose Yourself"
                        }
                    },
                    {
                        "name": "artist",
                        "in": "query",
                        "description": "Nom de l'artiste",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Eminem"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Résultats de la recherche"
                    },
                    "404": {
                        "description": "Aucune chanson trouvée"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "API Token",
                "scheme": "bearer"
            }
        }
    },
...
}